[24] Change install procedure :art:
