+++ positional_encodings/tf_positional_encodings.py
@@ -177,3 +177,26 @@ class TFPositionalEncoding3D(tf.keras.layers.Layer):
             emb[None, :, :, :, :org_channels], tf.shape(inputs)[0], axis=0
         )
         return self.cached_penc
+
+
+class TFSummer(tf.keras.layers.Layer):
+    def __init__(self, penc):
+        """
+        :param model: The type of positional encoding to run the summer on.
+        """
+        super(TFSummer, self).__init__()
+        self.penc = penc
+
+    @tf.function
+    def call(self, tensor):
+        """
+        :param tensor: A 3, 4 or 5d tensor that matches the model output size
+        :return: Positional Encoding Matrix summed to the original tensor
+        """
+        penc = self.penc(tensor)
+        assert (
+            tensor.shape == penc.shape
+        ), "The original tensor size {} and the positional encoding tensor size {} must match!".format(
+            tensor.size, penc.size
+        )
+        return tensor + penc
