+++ requirements.txt
@@ -25,7 +25,7 @@ idna==3.3
 importlib-metadata==4.10.0
 iniconfig==1.1.1
 isort==5.10.1
-keras==2.8.0
+keras==2.9.1
 Keras-Preprocessing==1.1.2
 libclang==12.0.0
 Markdown==3.3.6
@@ -55,11 +55,11 @@ requests==2.26.0
 requests-oauthlib==1.3.0
 rsa==4.8
 six==1.16.0
-tensorboard==2.8.0
+tensorboard==2.9.1
 tensorboard-data-server==0.6.1
 tensorboard-plugin-wit==1.8.0
-tensorflow==2.8.0
-tensorflow-estimator==2.8.0
+tensorflow==2.9.1
+tensorflow-estimator==2.9.1
 tensorflow-io-gcs-filesystem==0.23.1
 termcolor==1.1.0
 toml==0.10.2
